requirements.txt
altair==5.4.1
annotated-types==0.7.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
blis==0.7.11
bokeh==3.5.2
cachetools==5.5.0
catalogue==2.0.10
catboost==1.2.7
category-encoders==2.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
deepdiff==8.0.1
deprecation==2.1.0
dipy==1.9.0
eli5==0.13.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.9.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
graphviz==0.20.3
h5py==3.11.0
idna==3.8
imbalanced-learn==0.12.3
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lightgbm==4.5.0
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
narwhals==1.6.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
numba==0.60.0
numpy==1.26.4
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.1.4
pandas-datareader==0.10.0
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
pycaret==3.3.2
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.1
Pygments==2.18.0
pyod==2.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-time==0.3.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.8.0
rpds-py==0.20.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
setuptools==75.1.0
setuptools-scm==8.1.0
shellingham==1.5.4
six==1.16.0
sktime==0.26.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
stats==0.1.2a0
statsmodels==0.14.3
streamlit==1.38.0
sympy==1.13.2
tabulate==0.9.0
tbats==1.1.3
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
TIME-python==0.0.17
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trx-python==0.3
tsdownsample==0.1.3
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==2.1.1
XlsxWriter==3.2.0
xxhash==3.5.0
xyzservices==2024.9.0
yellowbrick==1.5
zipp==3.20.2
